clean plot tests
